Refactored TaskGroup.jsx
